requirements.txt
tqdm
nltk
PyEnchant
re